Conflicts:
	source/include/fds_commons_rm.h
	source/java/src/com/formationds/om/OmConfigurationApi.java
	source/java/src/com/formationds/om/webkit/rest/SetVolumeQosParams.java
	source/java/src/com/formationds/om/webkit/rest/domain/PutLocalDomain.java
	source/java/src/com/formationds/xdi/XdiConfigurationApi.java
	source/orch-mgr/om-dlt/OmResources.cpp
	source/orch-mgr/om-lib/OmReqHandler.cpp
	source/orch-mgr/om-lib/orchMgr.cpp
	source/orch-mgr/orch-cli/fdscli.cpp

+++ source/java/src/com/formationds/om/webkit/rest/CreateVolume.java
@@ -166,9 +166,9 @@ public class CreateVolume
           Thread.sleep( 200 );
           SetVolumeQosParams.setVolumeQos( configApi,
                                            volume.getName(),
-                                           ( int ) volume.getSla(),
+                                           volume.getSla(),
                                            volume.getPriority(),
-                                           ( int ) volume.getLimit(),
+                                           volume.getLimit(),
                                            volume.getCommit_log_retention(),
                                            volume.getMediaPolicy() );
